👌 move get ip to test utils
