--- scripts/get-ip.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-ifconfig | grep -Eo 'inet (addr:)?([0-9]*\.){3}[0-9]*' | grep -Eo '([0-9]*\.){3}[0-9]*' | grep -v '127.0.0.1' | head -n 1
