+++ test/e2e/wdio.bs.conf.js
@@ -1,11 +1,6 @@
-const execSync = require('child_process').execSync
-const path = require('path')
-
 const baseConf = require('./wdio.base.conf')
 const browsers = require('./browsers.conf')
-const getBuildInfos = require('../getBuildInfos')
-
-const ipAddress = execSync(path.join(__dirname, '../..', 'scripts/get-ip.sh'))
+const { getBuildInfos, getIp } = require('../utils')
 
 exports.config = {
   ...baseConf,
@@ -20,7 +15,7 @@ exports.config = {
   logLevels: {
     '@wdio/browserstack-service': 'info',
   },
-  baseUrl: `http://${ipAddress}:3000`,
+  baseUrl: `http://${getIp()}:3000`,
   services: ['browserstack'],
   browserstackLocal: true,
   user: process.env.BS_USERNAME,
