+++ test/unit/karma.bs.conf.js
@@ -1,9 +1,6 @@
-const execSync = require('child_process').execSync
-const path = require('path')
-
 const karmaBaseConf = require('./karma.base.conf')
 const browsers = require('../browsers.conf')
-const getBuildInfos = require('../getBuildInfos')
+const { getBuildInfos, getIp } = require('../utils')
 
 // force entry resolution to ensure sinon code is in ES5
 // https://github.com/webpack/webpack/issues/5756
@@ -11,7 +8,6 @@ const getBuildInfos = require('../getBuildInfos')
 karmaBaseConf.webpack.resolve.mainFields = ['cdn', 'main']
 
 const ONE_MINUTE = 60000
-const ipAddress = execSync(path.join(__dirname, '../..', 'scripts/get-ip.sh'))
 
 module.exports = function(config) {
   config.set({
@@ -20,7 +16,7 @@ module.exports = function(config) {
     reporters: [...karmaBaseConf.reporters, 'BrowserStack'],
     browsers: Object.keys(browsers),
     concurrency: 5,
-    hostname: ipAddress,
+    hostname: getIp(),
     captureTimeout: ONE_MINUTE,
     browserDisconnectTimeout: ONE_MINUTE,
     browserDisconnectTolerance: 3,
